requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
dateparser==0.7.1
discord.py==1.5.1
dj-database-url==0.5.0
Django==2.1.7
facebook-sdk==3.1.0
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
nltk==3.4
oauth2client==4.1.3
oauthlib==3.0.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyNaCl==1.2.1
PySocks==1.6.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
regex==2019.2.21
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
textblob==0.15.3
tweepy==3.7.0
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.24.2
websockets==6.0
whitenoise==4.1.2
yarl==1.3.0
